chat-with-markdown/requirements.txt
altair==5.1.2
attrs==23.1.0
black==23.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
filelock==3.12.4
fire==0.5.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
protobuf==4.24.4
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
streamlit==1.27.2
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
